chatgpt呼び出しのapi修正する
