+++ app/controllers/application_controller.rb
@@ -23,7 +23,7 @@
 end"""
 class ApplicationController < ActionController::Base
   #protect_from_forgery with: :null_session #追加
-  skip_before_action :verify_authenticity_token
+  #skip_before_action :verify_authenticity_token
   before_action :authenticate_request
 
   private
